src/ProtectManager/component/PM_System_Base_Common_Service/src/pm-common/src/test/java/openbackup/system/base/common/utils/DateFormatUtilTest.java
/*
* This file is a part of the open-eBackup project.
* This Source Code Form is subject to the terms of the Mozilla Public License, v. 2.0.
* If a copy of the MPL was not distributed with this file, You can obtain one at
* http://mozilla.org/MPL/2.0/.
*
* Copyright (c) [2024] Huawei Technologies Co.,Ltd.
*
* THIS SOFTWARE IS PROVIDED ON AN "AS IS" BASIS, WITHOUT WARRANTIES OF ANY KIND,
* EITHER EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO NON-INFRINGEMENT,
* MERCHANTABILITY OR FIT FOR A PARTICULAR PURPOSE.
*/
package openbackup.system.base.common.utils;

import openbackup.system.base.common.utils.DateFormatUtil;

import org.junit.Assert;
import org.junit.Test;

import java.text.DateFormat;
import java.text.ParseException;
import java.util.Date;
import java.util.Locale;

public class DateFormatUtilTest {
    @Test
    public void testFormat() {
        DateFormat dateFormat = DateFormat.getDateInstance(DateFormat.DEFAULT, Locale.ENGLISH);
        Date date = new Date(1234);
        Assert.assertEquals(DateFormatUtil.format(dateFormat, date), "Jan 1, 1970");
        Assert.assertEquals(DateFormatUtil.format(dateFormat, (Object) date), "Jan 1, 1970");
    }

    @Test(expected = ParseException.class)
    public void testFormatRaiseException() throws Exception {
        DateFormat dateFormat = DateFormat.getDateInstance(DateFormat.DEFAULT, Locale.ENGLISH);
        DateFormatUtil.parse(dateFormat, "2010-12-02 20:02:01");
    }
}
